requirements.txt
greenlet==1.1.3
joblib==1.2.0
kafka-python==2.0.2
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.41
threadpoolctl==3.1.0
